api/requirements.txt
fastapi
uvicorn[standard]
ulid-py
pydantic
aioboto3
